+++ forge/src/main/java/com/iamkaf/amber/platform/ForgeAmberEventSetup.java
@@ -8,6 +8,7 @@ import com.iamkaf.amber.api.event.v1.events.common.PlayerEvents;
 import com.iamkaf.amber.api.event.v1.events.common.client.ClientCommandEvents;
 import com.iamkaf.amber.api.keymapping.KeybindHelper;
 import com.iamkaf.amber.platform.services.IAmberEventSetup;
+import net.minecraft.client.Minecraft;
 import net.minecraft.world.InteractionResult;
 import net.minecraftforge.client.event.RegisterClientCommandsEvent;
 import net.minecraftforge.client.event.RegisterKeyMappingsEvent;
@@ -33,6 +34,7 @@ public class ForgeAmberEventSetup implements IAmberEventSetup {
     @Override
     public void registerClient() {
         RegisterClientCommandsEvent.BUS.addListener(EventHandlerClient::onCommandRegistration);
+        // mod bus events
         RegisterKeyMappingsEvent.getBus((BusGroup) AmberMod.getEventBus(Constants.MOD_ID))
                 .addListener(EventHandlerClient::onKeybindRegistration);
     }